+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Added `Unauthorized` response template
+- Added `ServiceUnavailable` response template
 
 ## [2.0.0]